🤖 fix: move openaiProModeAvailable to browser-safe models.ts
The round-2 fix imported providerOptions.ts from browser components, which
pulls node-only logging (node:path via constants/paths) into the renderer
bundle and broke Build/Linux, Smoke/Server, and E2E (macos):
  'join' is not exported by __vite-browser-external

Moved openaiProModeAvailable + resolveProviderOptionsNamespaceKey into
models.ts (already browser-safe) and re-exported from providerOptions.ts for
existing node-side consumers. Verified make build-renderer locally.
