+++ src/common/utils/ai/providerOptions.ts
@@ -11,7 +11,7 @@ import type { GoogleGenerativeAIProviderOptions } from "@ai-sdk/google";
 import type { OpenAIResponsesProviderOptions } from "@ai-sdk/openai";
 import type { JSONValue } from "@ai-sdk/provider";
 import type { XaiProviderOptions } from "@ai-sdk/xai";
-import { PROVIDER_DEFINITIONS, type ProviderName } from "@/common/constants/providers";
+import type { ProviderName } from "@/common/constants/providers";
 import type { ProvidersConfigMap } from "@/common/orpc/types";
 import type { MuxProviderOptions } from "@/common/types/providerOptions";
 import type { OpenAIReasoningMode, ThinkingLevel } from "@/common/types/thinking";
@@ -29,7 +29,16 @@ import { isGeminiFlashThinkingLevelModelName } from "@/common/utils/thinking/pol
 import { resolveModelForMetadata } from "@/common/utils/providers/modelEntries";
 import { log } from "@/node/services/log";
 import type { MuxMessage } from "@/common/types/message";
-import { getExplicitGatewayPrefix, normalizeToCanonical, supports1MContext } from "./models";
+import {
+  normalizeToCanonical,
+  resolveProviderOptionsNamespaceKey,
+  supports1MContext,
+} from "./models";
+
+// Re-export for existing consumers (aiService, providerModelFactory): the
+// implementation moved to the browser-safe models.ts because this module
+// imports node-only logging.
+export { openaiProModeAvailable, resolveProviderOptionsNamespaceKey } from "./models";
 
 /**
  * Request header used to override Anthropic's `output_config.effort` at the
@@ -87,24 +96,6 @@ const OPENAI_REASONING_SUMMARY_UNSUPPORTED_MODELS = new Set<string>([
 
 function supportsOpenAIReasoningSummary(modelName: string): boolean {
   return !OPENAI_REASONING_SUMMARY_UNSUPPORTED_MODELS.has(modelName);
-}
-
-export function resolveProviderOptionsNamespaceKey(
-  canonicalProviderName: string,
-  routeProvider?: ProviderName
-): string {
-  const routeDefinition = routeProvider ? PROVIDER_DEFINITIONS[routeProvider] : undefined;
-  if (
-    !routeProvider ||
-    routeProvider === canonicalProviderName ||
-    (routeDefinition != null &&
-      "passthrough" in routeDefinition &&
-      routeDefinition.passthrough === true)
-  ) {
-    return canonicalProviderName;
-  }
-
-  return routeProvider;
 }
 
 function resolveAnthropic1MCapabilityModel(
@@ -653,29 +644,4 @@ export function buildRequestHeaders(
   }
 
   return Object.keys(headers).length > 0 ? headers : undefined;
-}
-
-/**
- * Route-aware pro-mode availability for UI surfaces (PRO toggle, palette command).
- *
- * Mirrors the buildRequestHeaders wire gating: pro mode only reaches OpenAI when
- * the request flows through our OpenAI fetch wrapper (direct `openai:` or a
- * passthrough gateway like mux-gateway). Explicit non-passthrough gateway model
- * strings (e.g. `openrouter:openai/gpt-5.6-sol`, `github-copilot:gpt-5.6-sol`)
- * never emit the pro-mode header, so surfacing the toggle there would persist a
- * setting that can never affect the request.
- */
-export function openaiProModeAvailable(modelString: string): boolean {
-  if (!openaiSupportsProMode(modelString)) {
-    return false;
-  }
-
-  const normalized = normalizeToCanonical(modelString);
-  const [origin] = normalized.split(":", 2);
-  if (origin !== "openai") {
-    return false;
-  }
-
-  const explicitGateway = getExplicitGatewayPrefix(modelString);
-  return resolveProviderOptionsNamespaceKey(origin, explicitGateway) === origin;
 }
